backend/requirements.txt
fastapi
uvicorn
scikit-learn
catboost
joblib
requests
pandas
numpy
python-multipart
python-dotenv
